rename: dialog component
