+++ src/lib/components/Card/FacultyCard.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import * as Dialog from '$lib/components/ui/dialog';
+	import * as Dialog from '$lib/components/ui/dialog-nandanfucker';
 
 	export let image: string;
 	export let image2: string = image;
